fixed double vs float bug in rectification

+++ src/Frame.cpp
@@ -4,7 +4,7 @@
 CameraATAN Frame::cameraModel = CameraATAN();
 Detector   Frame::detector    = Detector();
 PreProc    Frame::preproc     = PreProc();
-long unsigned int Frame::idCounter   = 0;
-long unsigned int Frame::kfIdCounter   = 0;
+int Frame::idCounter   = 0;
+int Frame::kfIdCounter   = 0;
 float Frame::averageQuality = 0.8;
